requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
astunparse==1.6.3
attrs==25.1.0
banal==1.0.6
beautifulsoup4==4.13.4
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
coverage==7.8.0
cryptography==44.0.3
cycler==0.12.1
dill==0.3.8
dnspython==2.7.0
eventlet==0.39.1
filelock==3.18.0
firebase-admin==6.8.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
httplib2==0.22.0
huggingface-hub==0.31.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.8
kmodes==0.12.2
libclang==18.1.1
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
mysql-connector-python==9.2.0
namex==0.0.8
narwhals==1.29.1
ndg-httpsclient==0.5.1
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
p2j==1.3.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pusher==3.3.3
PyJWT==2.10.1
PyMySQL==1.1.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.40
streamlit==1.43.1
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
